Move all match templates to jinja2
