+++ controllers/match_controller.py
@@ -1,10 +1,9 @@
 import os
 
-from google.appengine.ext.webapp import template
-
 from base_controller import CacheableHandler
 from models.event import Event
 from models.match import Match
+from template_engine import jinja2_engine
 
 
 class MatchDetail(CacheableHandler):
@@ -52,5 +51,4 @@ class MatchDetail(CacheableHandler):
         if event.within_a_day:
             self._cache_expiration = self.SHORT_CACHE_EXPIRATION
 
-        path = os.path.join(os.path.dirname(__file__), '../templates/match_details.html')
-        return template.render(path, self.template_values)
+        return jinja2_engine.render('match_details.html', self.template_values)
