+++ template_engine/jinja2_filters.py
@@ -4,6 +4,24 @@ import time
 import urllib
 
 
+defense_render_names_2016 = {
+    'A_ChevalDeFrise': 'Cheval De Frise',
+    'A_Portcullis': 'Portcullis',
+    'B_Ramparts': 'Ramparts',
+    'B_Moat': 'Moat',
+    'C_SallyPort': 'Sally Port',
+    'C_Drawbridge': 'Drawbridge',
+    'D_RoughTerrain': 'Rough Terrain',
+    'D_RockWall': 'Rock Wall'
+}
+
+
+def defense_name(value):
+    if value in defense_render_names_2016:
+        return defense_render_names_2016[value]
+    return value
+
+
 def digits(s):
     if not s:
         return ''
