--- templates/match_partials/full_match_table.html
@@ -1,42 +1,0 @@
-<table class="match-table">
-  <thead>
-    {% include "match_partials/match_table_header.html" %}
-  </thead>
-
-  <tbody>
-    {% if matches.qm %}
-      <tr class="key"><th colspan="16">Qualifications</th></tr>
-      {% for match in matches.qm %}
-        {% include "match_partials/match_row.html" %}
-      {% endfor %}
-    {% endif %}
-
-    {% if matches.ef %}
-      <tr class="key"><th colspan="16">Eighth-Finals</th></tr>
-      {% for match in matches.ef %}
-        {% include "match_partials/match_row.html" %}
-      {% endfor %}
-    {% endif %}
-
-    {% if matches.qf %}
-      <tr class="key"><th colspan="16">Quarterfinals</th></tr>
-      {% for match in matches.qf %}
-        {% include "match_partials/match_row.html" %}
-      {% endfor %}
-    {% endif %}
-
-    {% if matches.sf %}
-      <tr class="key"><th colspan="16">Semifinals</th></tr>
-      {% for match in matches.sf %}
-        {% include "match_partials/match_row.html" %}
-      {% endfor %}
-    {% endif %}
-
-    {% if matches.f %}
-      <tr class="key"><th colspan="16">Finals</th></tr>
-      {% for match in matches.f %}
-        {% include "match_partials/match_row.html" %}
-      {% endfor %}
-    {% endif %}
-  </tbody>
-</table>
