--- templates/match_partials/match_table_header.html
@@ -1,18 +1,0 @@
-<tr class="key visible-lg">
-  <th><span class="glyphicon glyphicon-play-circle" rel="tooltip" title="Matches with the play icon have videos to watch"></span></th>
-  <th>Match</th>
-  <th colspan="6">Red Alliance</th>
-  <th colspan="6">Blue Alliance</th>
-  <th colspan="2">Scores</th>
-</tr>
-
-<tr class="key hidden-lg">
-  <th rowspan="2"><span class="glyphicon glyphicon-play-circle" title="Matches with the play icon have videos to watch"></span></th>
-  <th rowspan="2">Match</th>
-  <th colspan="12">Red Alliance</th>
-  <th colspan="2">Red Score</th>
-</tr>
-<tr class="key hidden-lg">
-  <th colspan="12">Blue Alliance</th>
-  <th colspan="2">Blue Score</th>
-</tr>
