--- templates/match_partials/match_table_upcoming_header.html
@@ -1,15 +1,0 @@
-<tr class="key visible-lg">
-  <th><span class="glyphicon glyphicon-time" rel="tooltip" title="Match times are approximate and are in your local timezone"></span></th>
-  <th>Match</th>
-  <th colspan="6">Red Alliance</th>
-  <th colspan="6">Blue Alliance</th>
-</tr>
-
-<tr class="key hidden-lg">
-  <th rowspan="2"><span class="glyphicon glyphicon-time" rel="tooltip" title="Match times are approximate and are in your local timezone"></span></th>
-  <th rowspan="2">Match</th>
-  <th colspan="12">Red Alliance</th>
-</tr>
-<tr class="key hidden-lg">
-  <th colspan="12">Blue Alliance</th>
-</tr>
