--- templates/match_partials/qual_match_table.html
@@ -1,16 +1,0 @@
-<table class="match-table">
-  <thead>
-    {% include "match_partials/match_table_header.html" %}
-  </thead>
-
-  <tbody>
-    {% if matches.qm %}
-      <tr class="key">
-        <th colspan="16">Qualifications</th>
-      </tr>
-      {% for match in matches.qm %}
-        {% include "match_partials/match_row.html" %}
-      {% endfor %}
-    {% endif %}
-  </tbody>
-</table>
