--- templates/match_partials/upcoming_match_table.html
@@ -1,11 +1,0 @@
-<table class="match-table">
-  <thead>
-    {% include "match_partials/match_table_upcoming_header.html" %}
-  </thead>
-
-  <tbody>
-    {% for match in matches %}
-      {% include "match_partials/match_upcoming_row.html" %}
-    {% endfor %}
-  </tbody>
-</table>
