+++ templates_jinja2/match_partials/match_breakdown_2015.html
@@ -59,9 +59,9 @@
     {% endif %}
 
     <tr class="key">
-      <td class="redScore">{{match.score_breakdown.red.tote_points|add:match.score_breakdown.red.container_points|add:match.score_breakdown.red.litter_points}}</td>
+      <td class="redScore">{{match.score_breakdown.red.tote_points + match.score_breakdown.red.container_points + match.score_breakdown.red.litter_points}}</td>
       <th>Total Teleop</th>
-      <td class="blueScore">{{match.score_breakdown.blue.tote_points|add:match.score_breakdown.blue.container_points|add:match.score_breakdown.blue.litter_points}}</td>
+      <td class="blueScore">{{match.score_breakdown.blue.tote_points + match.score_breakdown.blue.container_points + match.score_breakdown.blue.litter_points}}</td>
     </tr>
 
     {% if "coopertition_points" in match.score_breakdown %}
